+++ Classes/Domain/Model/Interfaces/CreationDate.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace OliverKlee\Oelib\Domain\Model\Interfaces;
+
+/**
+ * Interface for domain models with a creation date.
+ *
+ * The corresponding trait is the default implementation.
+ *
+ * @author Oliver Klee <typo3-coding@oliverklee.de
+ */
+interface CreationDate
+{
+    /**
+     * @return \DateTime|null
+     */
+    public function getCreationDate();
+
+    /**
+     * @param \DateTime $creationDate
+     *
+     * @return void
+     */
+    public function setCreationDate(\DateTime $creationDate);
+}